Sleep_Tracker_AI/sleepproject/pyproject.toml
[project]
name = "sleepproject"
version = "0.1.0"
description = ""
authors = [
    {name = "NeMoVetr",email = "131559606+NeMoVetr@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "amqp (>=5.3.1,<6.0.0)",
    "anyio (>=4.9.0,<5.0.0)",
    "asgiref (>=3.8.1,<4.0.0)",
    "async-timeout (>=5.0.1,<6.0.0)",
    "billiard (>=4.2.1,<5.0.0)",
    "build (>=1.2.2.post1,<2.0.0)",
    "cachecontrol (>=0.14.3,<0.15.0)",
    "celery (>=5.5.3,<6.0.0)",
    "certifi (>=2025.6.15,<2026.0.0)",
    "charset-normalizer (>=3.4.2,<4.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "click-didyoumean (>=0.3.1,<0.4.0)",
    "click-plugins (>=1.1.1,<2.0.0)",
    "click-repl (>=0.3.0,<0.4.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "distlib (>=0.3.9,<0.4.0)",
    "django (>=5.2.3,<6.0.0)",
    "dnspython (>=2.7.0,<3.0.0)",
    "dulwich (>=0.23.0,<0.24.0)",
    "eventlet (>=0.40.0,<0.41.0)",
    "fastjsonschema (>=2.21.1,<3.0.0)",
    "filelock (>=3.18.0,<4.0.0)",
    "findpython (>=0.6.3,<0.7.0)",
    "future (>=1.0.0,<2.0.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "h11 (>=0.16.0,<0.17.0)",
    "httpcore (>=1.0.9,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "idna (>=3.10,<4.0)",
    "installer (>=0.7.0,<0.8.0)",
    "jaraco-classes (>=3.4.0,<4.0.0)",
    "jaraco-context (>=6.0.1,<7.0.0)",
    "jaraco-functools (>=4.2.1,<5.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "keyring (>=25.6.0,<26.0.0)",
    "kombu (>=5.5.4,<6.0.0)",
    "markupsafe (>=3.0.2,<4.0.0)",
    "more-itertools (>=10.7.0,<11.0.0)",
    "msgpack (>=1.1.1,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "narwhals (>=1.44.0,<2.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "packaging (>=25.0,<26.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pbs-installer (>=2025.6.12,<2026.0.0)",
    "pip (>=25.1.1,<26.0.0)",
    "pip-review (>=1.3.0,<2.0.0)",
    "pkginfo (>=1.12.1.2,<2.0.0.0)",
    "platformdirs (>=4.3.8,<5.0.0)",
    "plotly (>=6.1.2,<7.0.0)",
    "prompt-toolkit (>=3.0.51,<4.0.0)",
    "pyproject-hooks (>=1.2.0,<2.0.0)",
    "pysocks (>=1.7.1,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-highcharts (>=0.4.2,<0.5.0)",
    "pytz (>=2025.2,<2026.0)",
    "pywin32-ctypes (>=0.2.3,<0.3.0)",
    "rapidfuzz (>=3.13.0,<4.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "requests-toolbelt (>=1.0.0,<2.0.0)",
    "shellingham (>=1.5.4,<2.0.0)",
    "six (>=1.17.0,<2.0.0)",
    "sniffio (>=1.3.1,<2.0.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "tomlkit (>=0.13.3,<0.14.0)",
    "trove-classifiers (>=2025.5.9.12,<2026.0.0.0)",
    "typing-extensions (>=4.14.0,<5.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "urllib3 (>=2.5.0,<3.0.0)",
    "vine (>=5.1.0,<6.0.0)",
    "virtualenv (>=20.31.2,<21.0.0)",
    "wcwidth (>=0.2.13,<0.3.0)",
    "wget (>=3.2,<4.0)",
    "zstandard (>=0.25.0,<0.26.0)",
    "celery-progress (>=0.5,<0.6)",
    "django-dragndrop-related (>=0.3.1,<0.4.0)",
    "django-debug-toolbar (>=6.1.0,<7.0.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)",
    "django-cursor-pagination (>=0.3.0,<0.4.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pylint (>=4.0.3,<5.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "coverage (>=7.12.0,<8.0.0)",
    "google-genai (>=1.52.0,<2.0.0)",
    "langfuse (>=3.10.1,<4.0.0)",
    "django-prometheus (>=2.4.1,<3.0.0)",
    "qdrant-client (>=1.16.1,<2.0.0)",
    "sentence-transformers (>=5.1.2,<6.0.0)",
    "pymupdf (>=1.26.6,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
